+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## NEXT
 
+### Added
+* **core:** dso-header logout link verrijken met logoutClick event ([#1605](https://github.com/dso-toolkit/dso-toolkit/issues/1605))
+
 ### Deprecated
 * **BREAKING: dso-toolkit:** Conclusion ([#1624](https://github.com/dso-toolkit/dso-toolkit/issues/1624))